integrated draft_exp in exp
