--- core/jobs/batch_jobs/exp_drafts_migration_job_test.py
@@ -1,378 +1,0 @@
-# coding: utf-8
-#
-# Copyright 2026 The Oppia Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS-IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Unit tests for exp_drafts_migration_job."""
-
-from __future__ import annotations
-
-from core import feconf
-from core.domain import (
-    draft_upgrade_services,
-    exp_domain,
-    rights_manager,
-    state_domain,
-)
-from core.jobs import job_test_utils
-from core.jobs.batch_jobs import exp_drafts_migration_job
-from core.jobs.types import job_run_result
-from core.platform import models
-from core.tests import test_utils
-
-MYPY = False
-if MYPY:  # pragma: no cover
-    from mypy_imports import exp_models, user_models
-
-(exp_models, user_models) = models.Registry.import_models(
-    [models.Names.EXPLORATION, models.Names.USER]
-)
-
-
-class MigrateExplorationDraftsJobTests(
-    job_test_utils.JobTestBase, test_utils.GenericTestBase
-):
-
-    JOB_CLASS = exp_drafts_migration_job.MigrateExplorationDraftsJob
-
-    EXP_ID = 'exp_1'
-    USER_ID = 'user_1'
-
-    def test_empty_storage(self) -> None:
-        self.assert_job_output_is_empty()
-
-    def test_migrates_draft_successfully(self) -> None:
-        exp_model = self.create_model(
-            exp_models.ExplorationModel,
-            id=self.EXP_ID,
-            title='title',
-            category='category',
-            language_code='en',
-            states_schema_version=feconf.CURRENT_STATE_SCHEMA_VERSION,
-            init_state_name='Introduction',
-            version=2,
-            states={
-                'Introduction': state_domain.State.create_default_state(
-                    'Introduction',
-                    content_id_for_state_content='content_0',
-                    content_id_for_default_outcome='default_outcome_1',
-                    is_initial_state=True,
-                ).to_dict()
-            },
-        )
-        rights_manager.create_new_exploration_rights(self.EXP_ID, self.USER_ID)
-        exp_model.commit(
-            self.USER_ID,
-            'Create exploration',
-            [{'cmd': exp_domain.CMD_CREATE_NEW}],
-        )
-
-        draft_id = '%s.%s' % (self.USER_ID, self.EXP_ID)
-        user_data_model = self.create_model(
-            user_models.ExplorationUserDataModel,
-            id=draft_id,
-            user_id=self.USER_ID,
-            exploration_id=self.EXP_ID,
-            draft_change_list=[
-                {
-                    'cmd': 'edit_state_property',
-                    'state_name': 'Introduction',
-                    'property_name': 'content',
-                    'new_value': {'html': 'Old content format'},
-                }
-            ],
-            draft_change_list_exp_version=1,
-            draft_change_list_last_updated=None,
-        )
-        user_data_model.update_timestamps()
-        user_data_model.put()
-
-        new_change_list = [{'cmd': 'upgraded_cmd'}]
-        with self.swap_to_always_return(
-            draft_upgrade_services,
-            'try_upgrading_draft_to_exp_version',
-            new_change_list,
-        ):
-            self.assert_job_output_is(
-                [
-                    job_run_result.JobRunResult(
-                        stdout='DRAFT PROCESSED SUCCESS: 1'
-                    )
-                ]
-            )
-
-        migrated_model = user_models.ExplorationUserDataModel.get(
-            self.USER_ID, self.EXP_ID
-        )
-
-        assert migrated_model is not None
-        self.assertEqual(migrated_model.draft_change_list, new_change_list)
-        self.assertEqual(migrated_model.draft_change_list_exp_version, 2)
-
-    def test_migrates_draft_updates_model_when_changes_detected(self) -> None:
-        self.save_new_valid_exploration(self.EXP_ID, self.USER_ID)
-
-        exp_model = exp_models.ExplorationModel.get(self.EXP_ID)
-        exp_model.version = 2
-        exp_model.update_timestamps()
-        exp_model.put()
-
-        old_draft_list = [{'cmd': 'old_cmd'}]
-        user_data_model = self.create_model(
-            user_models.ExplorationUserDataModel,
-            id='%s.%s' % (self.USER_ID, self.EXP_ID),
-            user_id=self.USER_ID,
-            exploration_id=self.EXP_ID,
-            draft_change_list=old_draft_list,
-            draft_change_list_exp_version=1,
-            draft_change_list_last_updated=None,
-        )
-        user_data_model.update_timestamps()
-        user_data_model.put()
-        new_change_list = [{'cmd': 'new_cmd'}]
-        with self.swap_to_always_return(
-            draft_upgrade_services,
-            'try_upgrading_draft_to_exp_version',
-            new_change_list,
-        ):
-            self.assert_job_output_is(
-                [
-                    job_run_result.JobRunResult(
-                        stdout='DRAFT PROCESSED SUCCESS: 1'
-                    )
-                ]
-            )
-
-        migrated_model = user_models.ExplorationUserDataModel.get(
-            self.USER_ID, self.EXP_ID
-        )
-
-        assert migrated_model is not None
-
-        self.assertEqual(migrated_model.draft_change_list, new_change_list)
-
-    def test_migrate_draft_with_empty_change_list_returns_ok(self) -> None:
-        exp_model = self.create_model(
-            exp_models.ExplorationModel,
-            id=self.EXP_ID,
-            title='title',
-            category='category',
-            language_code='en',
-            states_schema_version=feconf.CURRENT_STATE_SCHEMA_VERSION,
-            init_state_name='Introduction',
-            version=2,
-            states={
-                'Introduction': state_domain.State.create_default_state(
-                    'Introduction',
-                    content_id_for_state_content='content_0',
-                    content_id_for_default_outcome='default_outcome_1',
-                    is_initial_state=True,
-                ).to_dict()
-            },
-        )
-
-        user_model = self.create_model(
-            user_models.ExplorationUserDataModel,
-            id='%s.%s' % (self.USER_ID, self.EXP_ID),
-            user_id=self.USER_ID,
-            exploration_id=self.EXP_ID,
-            draft_change_list=[],
-            draft_change_list_exp_version=1,
-        )
-
-        result_item = exp_drafts_migration_job.MigrateExplorationDrafts._migrate_draft(  # pylint: disable=protected-access
-            user_model, exp_model
-        )
-
-        self.assertTrue(result_item.is_ok())
-        self.assertEqual(
-            result_item.unwrap(),
-            ('%s.%s' % (self.USER_ID, self.EXP_ID), user_model),
-        )
-
-    def test_migrate_draft_returns_error_on_failure(self) -> None:
-        exp_model = self.create_model(
-            exp_models.ExplorationModel,
-            id=self.EXP_ID,
-            title='title',
-            category='category',
-            language_code='en',
-            states_schema_version=feconf.CURRENT_STATE_SCHEMA_VERSION,
-            init_state_name='Introduction',
-            version=2,
-            states={},
-        )
-        user_model = self.create_model(
-            user_models.ExplorationUserDataModel,
-            id='%s.%s' % (self.USER_ID, self.EXP_ID),
-            user_id=self.USER_ID,
-            exploration_id=self.EXP_ID,
-            draft_change_list=[{'cmd': 'some_cmd'}],
-            draft_change_list_exp_version=1,
-        )
-
-        # Here we use object because we do not care about the argument types
-        # for this mock function, as it raises an exception regardless of input.
-        def mock_migrate_error(*args: object) -> None:
-            raise Exception('Migration failed')
-
-        with self.swap(
-            draft_upgrade_services,
-            'try_upgrading_draft_to_exp_version',
-            mock_migrate_error,
-        ):
-            result_item = exp_drafts_migration_job.MigrateExplorationDrafts._migrate_draft(  # pylint: disable=protected-access
-                user_model, exp_model
-            )
-
-        self.assertTrue(result_item.is_err())
-        self.assertIn('Migration failed', str(result_item.unwrap_err()[1]))
-
-    def test_migrate_draft_returns_error_for_incompatible_commits(self) -> None:
-        exp_model = self.create_model(
-            exp_models.ExplorationModel,
-            id=self.EXP_ID,
-            title='title',
-            category='category',
-            language_code='en',
-            states_schema_version=feconf.CURRENT_STATE_SCHEMA_VERSION,
-            init_state_name='Introduction',
-            version=2,
-            states={},
-        )
-        user_model = self.create_model(
-            user_models.ExplorationUserDataModel,
-            id='%s.%s' % (self.USER_ID, self.EXP_ID),
-            user_id=self.USER_ID,
-            exploration_id=self.EXP_ID,
-            draft_change_list=[{'cmd': 'some_cmd'}],
-            draft_change_list_exp_version=1,
-        )
-
-        with self.swap_to_always_return(
-            draft_upgrade_services,
-            'try_upgrading_draft_to_exp_version',
-            None,
-        ):
-            result_item = exp_drafts_migration_job.MigrateExplorationDrafts._migrate_draft(  # pylint: disable=protected-access
-                user_model, exp_model
-            )
-
-        self.assertTrue(result_item.is_err())
-        self.assertIn(
-            'due to incompatible commits', str(result_item.unwrap_err()[1])
-        )
-
-    def test_skips_users_without_drafts(self) -> None:
-        exp_model = self.create_model(
-            exp_models.ExplorationModel,
-            id=self.EXP_ID,
-            title='title',
-            category='category',
-            language_code='en',
-            states_schema_version=feconf.CURRENT_STATE_SCHEMA_VERSION,
-            init_state_name='Introduction',
-            version=2,
-            states={
-                'Introduction': state_domain.State.create_default_state(
-                    'Introduction',
-                    content_id_for_state_content='content_0',
-                    content_id_for_default_outcome='default_outcome_1',
-                    is_initial_state=True,
-                ).to_dict()
-            },
-        )
-        rights_manager.create_new_exploration_rights(self.EXP_ID, self.USER_ID)
-        exp_model.commit(
-            self.USER_ID,
-            'Create exploration',
-            [{'cmd': exp_domain.CMD_CREATE_NEW}],
-        )
-
-        no_draft_id = 'user_2.%s' % self.EXP_ID
-        no_draft_model = self.create_model(
-            user_models.ExplorationUserDataModel,
-            id=no_draft_id,
-            user_id='user_2',
-            exploration_id=self.EXP_ID,
-            draft_change_list=None,
-            draft_change_list_exp_version=1,
-        )
-        no_draft_model.update_timestamps()
-        no_draft_model.put()
-
-        self.assert_job_output_is_empty()
-
-
-class AuditMigrateExplorationDraftsJobTests(
-    job_test_utils.JobTestBase, test_utils.GenericTestBase
-):
-
-    JOB_CLASS = exp_drafts_migration_job.AuditMigrateExplorationDraftsJob
-
-    EXP_ID = 'exp_1'
-    USER_ID = 'user_1'
-
-    def test_empty_storage(self) -> None:
-        self.assert_job_output_is_empty()
-
-    def test_audits_draft_successfully_without_saving(self) -> None:
-        self.save_new_valid_exploration(self.EXP_ID, self.USER_ID)
-
-        exp_model = exp_models.ExplorationModel.get(self.EXP_ID)
-        exp_model.version = 2
-        exp_model.update_timestamps()
-        exp_model.put()
-
-        old_draft_list = [{'cmd': 'old_cmd'}]
-        user_data_model = self.create_model(
-            user_models.ExplorationUserDataModel,
-            id='%s.%s' % (self.USER_ID, self.EXP_ID),
-            user_id=self.USER_ID,
-            exploration_id=self.EXP_ID,
-            draft_change_list=old_draft_list,
-            draft_change_list_exp_version=1,
-            draft_change_list_last_updated=None,
-        )
-        user_data_model.update_timestamps()
-        user_data_model.put()
-
-        new_change_list = [{'cmd': 'new_cmd'}]
-        with self.swap_to_always_return(
-            draft_upgrade_services,
-            'try_upgrading_draft_to_exp_version',
-            new_change_list,
-        ):
-            self.assert_job_output_is(
-                [
-                    job_run_result.JobRunResult(
-                        stdout='DRAFT PROCESSED SUCCESS: 1'
-                    ),
-                    job_run_result.JobRunResult(
-                        stdout=(
-                            f'AUDIT MIGRATED DRAFT - User ID: {self.USER_ID}, '
-                            f'Exploration ID: {self.EXP_ID}, '
-                            f'Upgraded to Exp Version: 2'
-                        )
-                    ),
-                ]
-            )
-
-        unmigrated_model = user_models.ExplorationUserDataModel.get(
-            self.USER_ID, self.EXP_ID
-        )
-
-        assert unmigrated_model is not None
-        self.assertEqual(unmigrated_model.draft_change_list, old_draft_list)
-        self.assertEqual(unmigrated_model.draft_change_list_exp_version, 1)
