+++ scripts/backend_test_shards.json
@@ -204,8 +204,7 @@
         "core.jobs.batch_jobs.question_migration_jobs_test",
         "core.jobs.batch_jobs.subtopic_migration_jobs_test",
         "core.jobs.batch_jobs.reject_invalid_suggestion_and_delete_invalid_translation_jobs_test",
-        "core.domain.voiceover_cloud_task_services_test",
-        "core.jobs.batch_jobs.exp_drafts_migration_job_test"
+        "core.domain.voiceover_cloud_task_services_test"
     ],
     "4": [
         "core.storage.learner_group.gae_models_test",
